Properly implemented the resource selection buttons in the auction window. Added implementation of the buttons to increase or decrease the trade amount

+++ Game/Assets/Classes/Agents/Agent.cs
@@ -34,7 +34,7 @@ public abstract class Agent {
 	/// </summary>
 	/// <param name="type">The resource type.</param>
 	/// <exception cref="System.ArgumentException">If the specified type is <c>Data.ResourceType.NONE</c></exception>
-	public float GetResourceAmount(Data.ResourceType type) {
+	public int GetResourceAmount(Data.ResourceType type) {
 		switch (type) {
 			case Data.ResourceType.ENERGY:
 				return resources.GetEnergy();
